+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ## **WORK IN PROGRESS**
 - ...
 
+## v0.2.0 Nixos Integration 
+
+- docs: how to use the slack2zammad package and module
+- feature: systemd service nixos
+- feature: package nixos
+- feature: CLI arguments (-e, -h and -v)
+- feature: .env error handling
+
+
 ## v0.1.0 Initial working version
 
 - First official release 🎉